fixed dependency issues

+++ src/DataGPT/Infrastructure/AiClientConfig.cs
@@ -1,2 +1,2 @@
 ﻿namespace DataGPT.Net.Infrastructure;
-public record AiClientConfig(string OrganizationId, string SecreteKey, int NumberOfRetries, int Variance);
+public record AiClientConfig(string OrganizationId, string SecreteKey, int NumberOfRetries, double Variance);
